# docs/v1/swagger.yaml
basePath: /v1
definitions:
  mp.AppModelParams:
    properties:
      providerOpenAICompatible:
        allOf:
        - $ref: '#/definitions/mp.AppModelParamsOpenAICompatible'
        description: OpenAI-API-compatible模型配置
      providerYuanjing:
        allOf:
        - $ref: '#/definitions/mp.AppModelParamsYuanjing'
        description: YuanJing模型配置
    type: object
  mp.AppModelParamsOpenAICompatible:
    properties:
      llm:
        allOf:
        - $ref: '#/definitions/mp_openai_compatible.LLMParams'
        description: 大语言模型配置
    type: object
  mp.AppModelParamsYuanjing:
    properties:
      llm:
        allOf:
        - $ref: '#/definitions/mp_yuanjing.LLMParams'
        description: 大语言模型配置
    type: object
...
  mp_openai_compatible.LLMParams:
    properties:
      frequencyPenalty:
        description: 频率惩罚
        type: number
      frequencyPenaltyEnable:
        description: 频率惩罚(开关)
        type: boolean
      maxTokens:
        description: 最大标记
        type: integer
      maxTokensEnable:
        description: 最大标记(开关)
        type: boolean
      presencePenalty:
        description: 存在惩罚
        type: number
      presencePenaltyEnable:
        description: 存在惩罚(开关)
        type: boolean
      temperature:
        description: 温度
        type: number
      temperatureEnable:
        description: 温度(开关)
        type: boolean
      topP:
        description: Top P
        type: number
      topPEnable:
        description: Top P(开关)
        type: boolean
    type: object
...
  mp_yuanjing.LLMParams:
    properties:
      frequencyPenalty:
        description: 频率惩罚
        type: number
      frequencyPenaltyEnable:
        description: 频率惩罚(开关)
        type: boolean
      maxTokens:
        description: 最大标记
        type: integer
      maxTokensEnable:
        description: 最大标记(开关)
        type: boolean
      presencePenalty:
        description: 存在惩罚
        type: number
      presencePenaltyEnable:
        description: 存在惩罚(开关)
        type: boolean
      temperature:
        description: 温度
        type: number
      temperatureEnable:
        description: 温度(开关)
        type: boolean
      topP:
        description: Top P
        type: number
      topPEnable:
        description: Top P(开关)
        type: boolean
    type: object
...
  request.AppBriefConfig:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    required:
    - name
    type: object
  request.AppKnowledgeBase:
    properties:
      id:
        description: 知识库id
        type: string
      name:
        description: 知识库名称(请求非必填)
        type: string
    required:
    - id
    type: object
  request.AppKnowledgebaseConfig:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/request.AppKnowledgebaseParams'
        description: 知识库参数
      knowledgebases:
        description: 知识库id、名字
        items:
          $ref: '#/definitions/request.AppKnowledgeBase'
        type: array
    type: object
  request.AppKnowledgebaseParams:
    properties:
      maxHistory:
        description: 最长上下文
        type: integer
      maxHistoryEnable:
        description: 最长上下文(开关)
        type: boolean
      threshold:
        description: 过滤阈值
        type: number
      thresholdEnable:
        description: 过滤阈值(开关)
        type: boolean
      topK:
        description: 知识条数
        type: integer
      topKEnable:
        description: 知识条数(开关)
        type: boolean
    type: object
  request.AppModelConfig:
    properties:
      config:
        description: 模型配置
      displayName:
        description: 模型展示名称(请求非必填)
        type: string
      examples:
        allOf:
        - $ref: '#/definitions/mp.AppModelParams'
        description: 仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数
      model:
        description: 模型名称
        type: string
      modelId:
        description: 模型ID
        type: string
      modelType:
        description: 模型类型(llm/embedding/rerank)
        type: string
      provider:
        description: 模型供应商
        type: string
    type: object
...
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.ChatRagRequest:
    properties:
      question:
        type: string
      ragId:
        type: string
    required:
    - question
    - ragId
    type: object
  request.CheckFileListReq:
    properties:
      chunkName:
        description: 上传批次标识
        type: string
    required:
    - chunkName
    type: object
  request.CheckFileReq:
    properties:
      chunkName:
        description: 上传批次标识
        type: string
      fileName:
        description: 原始文件名
        type: string
      sequence:
        description: 分片文件序号
        type: integer
    required:
    - chunkName
    - fileName
    type: object
  request.CleanFileReq:
    properties:
      chunkName:
        description: 上传批次标识
        type: string
    required:
    - chunkName
    type: object
...
  request.DelApiKeyRequest:
    properties:
      apiId:
        description: ApiID
        type: string
    required:
    - apiId
    type: object
  request.DeleteAppSpaceAppRequest:
    properties:
      appId:
        description: 应用ID
        type: string
      appType:
        description: 应用类型
        type: string
    required:
    - appId
    - appType
    type: object
...
  request.DeleteFileReq:
    properties:
      fileList:
        description: 文件列表
        items:
          type: string
        type: array
      isExpired:
        description: minio存储文件是否过期 0:过期，1:不过期
        type: boolean
    type: object
...
  request.GenApiKeyRequest:
    properties:
      appId:
        description: 应用id
        type: string
      appType:
        description: 应用类型
        type: string
    required:
    - appId
    - appType
    type: object
...
  request.MergeFileReq:
    properties:
      chunkName:
        description: 上传批次标识
        type: string
      chunkTotal:
        description: 分片总数
        type: integer
      fileName:
        description: 原始文件名
        type: string
      fileSize:
        description: 原始文件大小
        type: integer
      isExpired:
        description: minio存储文件是否过期 0:过期，1:不过期
        type: boolean
    required:
    - chunkName
    - fileName
    type: object
...
  request.PublishAppRequest:
    properties:
      appId:
        description: 应用ID
        type: string
      appType:
        description: 应用类型
        type: string
      publishType:
        description: 发布类型(public:公开发布,private:私密发布)
        type: string
    type: object
...
  request.RagBrief:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      ragId:
        type: string
    required:
    - name
    - ragId
    type: object
  request.RagConfig:
    properties:
      knowledgeBaseConfig:
        allOf:
        - $ref: '#/definitions/request.AppKnowledgebaseConfig'
        description: 知识库
      modelConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: 模型
      ragId:
        type: string
      rerankConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: Rerank模型
    required:
    - knowledgeBaseConfig
    - modelConfig
    - ragId
    - rerankConfig
    type: object
  request.RagReq:
    properties:
      ragId:
        type: string
    required:
    - ragId
    type: object
...
  response.ApiResponse:
    properties:
      apiId:
        description: ApiID
        type: string
      apiKey:
        description: 生成的ApiKey
        type: string
      createdAt:
        description: 创建ApiKey的时间
        type: string
    type: object
  response.AppBriefInfo:
    properties:
      appId:
        description: 应用id
        type: string
      appType:
        description: 应用类型
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 应用图标
      createdAt:
        description: 应用创建时间
        type: string
      desc:
        description: 应用描述
        type: string
      name:
        description: 应用名称
        type: string
      publishType:
        description: 发布类型(public:公开发布,private:私密发布)
        type: string
      updatedAt:
        description: 应用更新时间(用于历史记录排序)
        type: string
    type: object
...
  response.CheckFileListResp:
    properties:
      uploadedFileSequences:
        description: 已经上传成功的切片文件序号列表
        items:
          type: integer
        type: array
    type: object
  response.CheckFileResp:
    properties:
      status:
        description: 0:不存在，1：已完成
        type: integer
    type: object
  response.CleanFileResp:
    properties:
      status:
        description: 0:清除失败，1：已完成
        type: integer
    type: object
...
  response.DeleteFileResp:
    properties:
      status:
        description: 0:删除失败，1：已完成
        type: integer
    type: object
...
  response.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  response.MergeFileResp:
    properties:
      fileName:
        description: 合并后文件名
        type: string
      filePath:
        description: minio文件的完整路径
        type: string
    type: object
...
  response.RagInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      knowledgeBaseConfig:
        allOf:
        - $ref: '#/definitions/request.AppKnowledgebaseConfig'
        description: 知识库
      modelConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: 模型
      name:
        description: 名称
        type: string
      ragId:
        type: string
      rerankConfig:
        allOf:
        - $ref: '#/definitions/request.AppModelConfig'
        description: Rerank模型
    required:
    - knowledgeBaseConfig
    - modelConfig
    - name
    - ragId
    - rerankConfig
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.UploadFileResp:
    properties:
      status:
        description: 0:上传失败，1：上传成功
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    ## HTTP Header
    | Header        | 说明      |
    | ------------- | --------- |
    | Authorization | JWT token |
    | X-Language    | 语言Code  |
    | X-Org-Id      | 组织ID    |

    ## HTTP Status
    | HTTP Status             | 说明                   |
    | ----------------------- | ---------------------- |
    | 200, StatusOK           | 请求返回成功           |
    | 400, StatusBadRequest   | 请求返回失败，用于业务 |
    | 401, StatusUnauthorized | JWT认证失败            |
    | 403, StatusForbidden    | 没有权限               |

    ## 权限-菜单对应表
    | 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |
    |-------------|-------|------|------|------|------|
    | guest       |       |      | 【访客】 |      |      |
    | common      |       |      | 【通用】 |      |      |
    | permission  |       |      | 权限管理 |      |      |
    | permission  | user  |      | 权限管理 | 用户管理 |      |
    | permission  | org   |      | 权限管理 | 组织管理 |      |
    | permission  | role  |      | 权限管理 | 角色管理 |      |

    ## `/v1/user/permission`返回用例
    ```json
    {
      "code": 0,
      "data": {
        "orgPermission": {
          "org": {"id": "test-org-id", "name": "test-org-name"},
          "permissions": [
            {"perm": "permission"},
            {"perm": "permission.user"},
            {"perm": "permission.org"},
            {"perm": "permission.role"}
          ]
        }
      },
      "msg": "操作成功"
    }
    ```
...
  version: v0.0.1
paths:
  /appspace/app:
    delete:
      consumes:
      - application/json
      description: 刪除智能体、工作流、文本问答等应用
      parameters:
      - description: 删除应用空间App参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteAppSpaceAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 刪除应用
      tags:
      - workspace.appspace
  /appspace/app/key:
    delete:
      consumes:
      - application/json
      description: 删除ApiKey
      parameters:
      - description: 删除Apikey参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DelApiKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除ApiKey
      tags:
      - workspace.appspace
    post:
      consumes:
      - application/json
      description: 生成ApiKey
      parameters:
      - description: 生成ApiKey参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GenApiKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ApiResponse'
              type: object
      summary: 生成ApiKey
      tags:
      - workspace.appspace
  /appspace/app/key/list:
    get:
      consumes:
      - application/json
      description: 获取ApiKey
      parameters:
      - description: 应用id
        in: query
        name: appId
        required: true
        type: string
      - description: 应用类型
        in: query
        name: appType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.ApiResponse'
                  type: array
              type: object
      summary: 获取ApiKey
      tags:
      - workspace.appspace
  /appspace/app/list:
    get:
      consumes:
      - application/json
      description: 获取智能体、工作流、文本问答等应用
      parameters:
      - description: 应用名(模糊查询)
        in: query
        name: name
        type: string
      - description: 应用类型 Enums(agent,workflow,rag)
        enum:
        - agent
        - workflow
        - rag
        in: query
        name: appType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.AppBriefInfo'
                        type: array
                    type: object
              type: object
      security:
      - JWT: []
      summary: 获取应用列表
      tags:
      - workspace.appspace
  /appspace/app/publish:
    post:
      consumes:
      - application/json
      description: 发布应用
      parameters:
      - description: 发布应用参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PublishAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 发布应用
      tags:
      - workspace.appspace
  /appspace/app/url:
    get:
      consumes:
      - application/json
      description: 获取Api根地址
      parameters:
      - description: 应用ID
        in: query
        name: appId
        required: true
        type: string
      - description: 应用类型
        in: query
        name: appType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: 获取Api根地址
      tags:
      - workspace.appspace
  /appspace/rag:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除RAG的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除RAG
      tags:
      - workspace.appspace.rag
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: ragId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.RagInfo'
              type: object
      security:
      - JWT: []
      summary: 获取RAG信息
      tags:
      - workspace.appspace.rag
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建RAG的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AppBriefConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/request.RagReq'
              type: object
      security:
      - JWT: []
      summary: 创建RAG
      tags:
      - workspace.appspace.rag
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新RAG基本信息的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RagBrief'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新RAG基本信息
      tags:
      - workspace.appspace.rag
  /appspace/rag/config:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新RAG配置信息的请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RagConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新RAG配置信息
      tags:
      - workspace.appspace.rag
...
  /file/check:
    get:
      consumes:
      - application/json
      description: 校验分片文件
      parameters:
      - description: 文件校验参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CheckFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CheckFileResp'
              type: object
      security:
      - JWT: []
      summary: 文件校验
      tags:
      - common
  /file/check/list:
    get:
      consumes:
      - application/json
      description: 校验分片文件列表
      parameters:
      - description: 文件列表校验参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CheckFileListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CheckFileListResp'
              type: object
      security:
      - JWT: []
      summary: 文件列表校验
      tags:
      - common
  /file/clean:
    post:
      consumes:
      - application/json
      description: 清除已上传的分片文件
      parameters:
      - description: 文件清除参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CleanFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CleanFileResp'
              type: object
      security:
      - JWT: []
      summary: 文件清除
      tags:
      - common
  /file/delete:
    delete:
      consumes:
      - application/json
      description: 删除已上传的文件
      parameters:
      - description: 文件删除请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DeleteFileResp'
              type: object
      security:
      - JWT: []
      summary: 文件删除
      tags:
      - common
  /file/merge:
    post:
      consumes:
      - application/json
      description: 合并分片文件，并上传minio
      parameters:
      - description: 文件合并参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MergeFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MergeFileResp'
              type: object
      security:
      - JWT: []
      summary: 文件合并
      tags:
      - common
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 分片文件上传
      parameters:
      - description: 原始文件名
        in: formData
        name: fileName
        required: true
        type: string
      - description: 分片文件序号
        in: formData
        name: sequence
        required: true
        type: integer
      - description: 上传批次标识
        in: formData
        name: chunkName
        required: true
        type: string
      - description: 文件
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UploadFileResp'
              type: object
      security:
      - JWT: []
      summary: 文件上传
      tags:
      - common
...
  /rag/chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: RAG问答请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ChatRagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 私域 RAG 问答
      tags:
      - workspace.appspace.rag
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"